Simplify examples
