+++ nextjs/app/transifex/basic-example/page.tsx
@@ -1,18 +1,12 @@
 'use client';
 
-import Link from 'next/link';
-import React from 'react';
 import { useLocale } from '@transifex/react';
 import { tx } from '@transifex/native';
-import {
-  ActionButton,
-  GreetingCard,
-  NotificationBadge,
-  ProductCard,
-} from './components';
+import Link from 'next/link';
+import { BasicTranslation, VariableExample, PluralExample } from './components';
 import '../styles.css';
 
-export default function ComponentsExamplePage() {
+export default function TransifexExample() {
   const locale = useLocale();
   const currentLocale = String(locale);
 
@@ -29,7 +23,7 @@ export default function ComponentsExamplePage() {
   return (
     <div className="container">
       <div className="header">
-        <h1 className="heading-1">Reusable Components</h1>
+        <h1 className="heading-1">Transifex Example</h1>
         <div className="lang-switcher">
           {languages.map((lang) => (
             <button
@@ -48,24 +42,9 @@ export default function ComponentsExamplePage() {
       </p>
 
       <div className="examples">
-        <GreetingCard name="Maria" />
-
-        <div className="notification-badges">
-          <NotificationBadge count={1} />
-          <NotificationBadge count={5} />
-          <NotificationBadge count={10} />
-        </div>
-
-        <div className="action-buttons">
-          <ActionButton action="Read" />
-          <ActionButton action="Close" />
-          <ActionButton action="Present" />
-        </div>
-
-        <div className="product-cards">
-          <ProductCard item="laptop" price="999.99" inStock={true} />
-          <ProductCard item="phone" price="599.99" inStock={false} />
-        </div>
+        <BasicTranslation />
+        <VariableExample />
+        <PluralExample />
       </div>
 
       <Link href="/transifex" className="link">
