+++ nextjs/app/transifex/styles.css
@@ -604,3 +604,426 @@
     color: rgb(147 197 253);
   }
 }
+
+/* Layout Components */
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 2rem;
+}
+
+.lang-switcher {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.debug-info {
+  font-size: 0.875rem;
+  color: rgb(75 85 99);
+  margin-bottom: 2rem;
+}
+
+@media (prefers-color-scheme: dark) {
+  .debug-info {
+    color: rgb(156 163 175);
+  }
+}
+
+.examples {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+}
+
+.notification-example,
+.action-buttons,
+.product-cards {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.notification-badges {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.75rem;
+}
+
+.counter-controls {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.action-buttons {
+  flex-direction: row;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
+
+.product-cards {
+  display: grid;
+  gap: 1rem;
+  grid-template-columns: repeat(auto-fit, minmax(250px, 1fr));
+}
+
+.back-links {
+  display: flex;
+  gap: 1rem;
+  margin-top: 2rem;
+}
+
+/* Button List */
+.button-list {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  margin-top: 2rem;
+  margin-bottom: 2rem;
+  max-width: 24rem;
+}
+
+.btn-primary {
+  display: block;
+  padding: 0.75rem 1.5rem;
+  background-color: rgb(37 99 235);
+  color: white;
+  border-radius: 0.5rem;
+  text-align: center;
+  text-decoration: none;
+  font-weight: 500;
+  transition: background-color 0.2s;
+}
+
+.btn-primary:hover {
+  background-color: rgb(29 78 216);
+}
+
+@media (prefers-color-scheme: dark) {
+  .btn-primary {
+    background-color: rgb(59 130 246);
+  }
+  .btn-primary:hover {
+    background-color: rgb(37 99 235);
+  }
+}
+
+/* Greeting Card */
+.greeting-card {
+  padding: 1.5rem;
+  border: 1px solid rgb(229 231 235);
+  border-radius: 0.5rem;
+  background-color: white;
+}
+
+@media (prefers-color-scheme: dark) {
+  .greeting-card {
+    border-color: rgb(55 65 81);
+    background-color: rgb(31 41 55);
+  }
+}
+
+.greeting-card-title {
+  font-size: 1.125rem;
+  font-weight: 600;
+  margin-bottom: 0.5rem;
+}
+
+.greeting-card-text {
+  font-size: 0.875rem;
+}
+
+/* Notification Badge */
+.notification-indicator {
+  position: relative;
+  display: flex;
+  height: 0.75rem;
+  width: 0.75rem;
+}
+
+.notification-ping {
+  position: absolute;
+  display: inline-flex;
+  height: 100%;
+  width: 100%;
+  border-radius: 9999px;
+  background-color: rgb(248 113 113);
+  opacity: 0.75;
+  animation: ping 1s cubic-bezier(0, 0, 0.2, 1) infinite;
+}
+
+@keyframes ping {
+  75%, 100% {
+    transform: scale(2);
+    opacity: 0;
+  }
+}
+
+.notification-dot {
+  position: relative;
+  display: inline-flex;
+  border-radius: 9999px;
+  height: 0.75rem;
+  width: 0.75rem;
+  background-color: rgb(239 68 68);
+}
+
+/* Social Card */
+.social-header-content {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.social-user-info {
+  display: flex;
+  align-items: center;
+  gap: 0.75rem;
+}
+
+.user-avatar {
+  height: 2.5rem;
+  width: 2.5rem;
+  border-radius: 9999px;
+  background: linear-gradient(to bottom right, rgb(96 165 250), rgb(37 99 235));
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: white;
+  font-weight: 600;
+}
+
+.post-meta {
+  font-size: 0.75rem;
+  color: rgb(107 114 128);
+}
+
+@media (prefers-color-scheme: dark) {
+  .post-meta {
+    color: rgb(156 163 175);
+  }
+}
+
+.delete-btn {
+  color: rgb(156 163 175);
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 0.25rem;
+}
+
+.delete-btn:hover {
+  color: rgb(220 38 38);
+}
+
+@media (prefers-color-scheme: dark) {
+  .delete-btn:hover {
+    color: rgb(248 113 113);
+  }
+}
+
+.post-content {
+  color: rgb(31 41 55);
+}
+
+@media (prefers-color-scheme: dark) {
+  .post-content {
+    color: rgb(229 231 235);
+  }
+}
+
+.action-buttons {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.icon {
+  height: 1.25rem;
+  width: 1.25rem;
+}
+
+/* Delete Modal */
+.delete-modal {
+  border-top: 1px solid rgb(229 231 235);
+  background-color: rgb(254 242 242);
+  padding: 1rem;
+}
+
+@media (prefers-color-scheme: dark) {
+  .delete-modal {
+    border-color: rgb(55 65 81);
+    background-color: rgba(127, 29, 29, 0.1);
+  }
+}
+
+.delete-message {
+  font-size: 0.875rem;
+  color: rgb(127 29 29);
+  margin-bottom: 0.75rem;
+}
+
+@media (prefers-color-scheme: dark) {
+  .delete-message {
+    color: rgb(254 242 242);
+  }
+}
+
+.loading-state {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0.5rem;
+}
+
+.modal-actions {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.btn-delete {
+  flex: 1;
+  border-radius: 0.375rem;
+  background-color: rgb(220 38 38);
+  padding: 0.5rem 1rem;
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: white;
+  border: none;
+  cursor: pointer;
+}
+
+.btn-delete:hover {
+  background-color: rgb(185 28 28);
+}
+
+.btn-cancel {
+  flex: 1;
+  border-radius: 0.375rem;
+  background-color: rgb(229 231 235);
+  padding: 0.5rem 1rem;
+  font-size: 0.875rem;
+  font-weight: 500;
+  border: none;
+  cursor: pointer;
+}
+
+.btn-cancel:hover {
+  background-color: rgb(209 213 219);
+}
+
+@media (prefers-color-scheme: dark) {
+  .btn-cancel {
+    background-color: rgb(55 65 81);
+    color: white;
+  }
+  .btn-cancel:hover {
+    background-color: rgb(75 85 99);
+  }
+}
+
+/* Upgrade Prompt */
+.upgrade-prompt {
+  border-top: 1px solid rgb(229 231 235);
+  background: linear-gradient(to right, rgb(243 232 255), rgb(219 234 254));
+  padding: 0.75rem;
+}
+
+@media (prefers-color-scheme: dark) {
+  .upgrade-prompt {
+    border-color: rgb(55 65 81);
+    background: linear-gradient(to right, rgba(88, 28, 135, 0.2), rgba(30, 64, 175, 0.2));
+  }
+}
+
+.upgrade-content {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+
+.upgrade-text {
+  font-size: 0.875rem;
+  color: rgb(55 65 81);
+}
+
+@media (prefers-color-scheme: dark) {
+  .upgrade-text {
+    color: rgb(209 213 219);
+  }
+}
+
+.btn-upgrade {
+  border-radius: 0.375rem;
+  background: linear-gradient(to right, rgb(147 51 234), rgb(37 99 235));
+  padding: 0.375rem 1rem;
+  font-size: 0.875rem;
+  font-weight: 500;
+  color: white;
+  border: none;
+  cursor: pointer;
+}
+
+.btn-upgrade:hover {
+  background: linear-gradient(to right, rgb(126 34 206), rgb(29 78 216));
+}
+
+/* Actions Showcase */
+.actions-showcase {
+  border-top: 1px solid rgb(229 231 235);
+  padding: 1rem;
+  background-color: rgb(249 250 251);
+}
+
+@media (prefers-color-scheme: dark) {
+  .actions-showcase {
+    border-color: rgb(55 65 81);
+    background-color: rgba(17, 24, 39, 0.5);
+  }
+}
+
+.showcase-label {
+  font-size: 0.75rem;
+  color: rgb(107 114 128);
+  margin-bottom: 0.5rem;
+}
+
+@media (prefers-color-scheme: dark) {
+  .showcase-label {
+    color: rgb(156 163 175);
+  }
+}
+
+.showcase-buttons {
+  display: flex;
+  gap: 0.5rem;
+}
+
+.btn-showcase-post {
+  border-radius: 0.25rem;
+  background-color: rgb(37 99 235);
+  padding: 0.25rem 0.75rem;
+  font-size: 0.75rem;
+  color: white;
+  border: none;
+  cursor: pointer;
+}
+
+.btn-showcase-post:hover {
+  background-color: rgb(29 78 216);
+}
+
+.btn-showcase-save {
+  border-radius: 0.25rem;
+  background-color: rgb(22 163 74);
+  padding: 0.25rem 0.75rem;
+  font-size: 0.75rem;
+  color: white;
+  border: none;
+  cursor: pointer;
+}
+
+.btn-showcase-save:hover {
+  background-color: rgb(21 128 61);
+}
